+++ python/lsst/obs/fiberSpectrograph/__init__.py
@@ -1,0 +1,3 @@
+from ._ingest import *
+from ._instrument import *
+from .spectrum import *
