+++ python/lsst/obs/fiberSpectrograph/filters.py
@@ -1,5 +1,5 @@
 from lsst.obs.base import FilterDefinition, FilterDefinitionCollection
 
-RUBIN_GENERIC_CAMERA_FILTER_DEFINITIONS = FilterDefinitionCollection(
+FIBER_SPECTROGRAPH_FILTER_DEFINITIONS = FilterDefinitionCollection(
     FilterDefinition(band="empty", physical_filter="empty"),
 )
