--- python/lsst/obs/rubinGenericCamera/__init__.py
@@ -1,1 +1,0 @@
-from ._instrument import *
