+++ setup.cfg
@@ -1,0 +1,5 @@
+[flake8]
+max-line-length = 110
+max-doc-length = 79
+ignore = W503, E203, N802, N803, N806, N812, N815, N816
+exclude = __init__.py
