+++ tests/test_instrument.py
@@ -1,10 +1,10 @@
-"""Tests of the RubinGenericCamera instrument class.
+"""Tests of the FiberSpectrograph instrument class.
 """
 
 import unittest
 
 import lsst.utils.tests
-import lsst.obs.rubinGenericCamera
+import lsst.obs.fiberSpectrograph
 from lsst.obs.base.instrument_tests import InstrumentTests, InstrumentTestData
 
 
@@ -12,11 +12,11 @@ class TestStarTracker(InstrumentTests, lsst.utils.tests.TestCase):
     def setUp(self):
         physical_filters = set(["empty"])
 
-        self.data = InstrumentTestData(name="StarTracker",
+        self.data = InstrumentTestData(name="FiberSpec",
                                        nDetectors=1,
                                        firstDetectorName="0",
                                        physical_filters=physical_filters)
-        self.instrument = lsst.obs.rubinGenericCamera.StarTracker()
+        self.instrument = lsst.obs.fiberSpectrograph.FiberSpectrograph()
 
 if __name__ == '__main__':
     lsst.utils.tests.init()
